Downgrade rust-toolchain to 1.43.1
Until the clippy bug generating a ton of warnings makes it to stable:
rust-lang/rust-clippy#5535

Also pin clippy checks in CI for the same rust version as the one
specified in the project rust-toolchain file

---

Fix test issues seen with 1.44 pre-emptively. We should also pin rust
versions on those CI runs now (will be done as part of a separate
cleanup).