Disable static linking of openssl for linux-gnu release build
We get the benefits of dynamic linking but since we are trying to
utilize the same binary as much as we can across packages, the
binary made on centos7 no longer works on debian based systems.

We will need to revisit debian packaging strategy here and may need
to support specific builds per package format.

Also add back the builder image based on debian for use as needed.