Make EventPayload lifetime laundering more robust
This tries to address 2 issues:

- No correlation between 'iref and 'head
- 'head is 'static

The first one is problematic because we know 'head and 'iref are very much
correlated lifetimes. If one could take a reference 'head that outlives 'iref,
that would be unsound.

The second issue is related and maybe even more problematic. While there might
not be any code patterns allowing the abuse of this wrongly marked 'static
lifetime. Upcoming features like Fn Traits and specialization could easily break
the assumptions here. More pressing even seem potential compiler optimization,
around 'will be alive for whole program duration' lifetimes.

To make the lie about 'static lifetimes more robust, we .. lie even more. By
transmuting the dependent lifetime from 'static to a lifetime bound to the
function call, we ensure that nothing can outlive the scope of the function we
know must be smaller than the lifetime of the object itself. In essence it's the
reverse lie found in the constructor, where we take a function local reference
lifetime and turn it into a 'static lifetime. Technically this approach is still
UB, but the assumption is that it has less potential of breaking in the future,
should for example the constrains of HRTBs be relaxed.

Signed-off-by: Lukas Bergdoll <lukas.bergdoll@gmail.com>