Artefact/Instance Lifecycle Improvements Part I.
For supporting two-phase startup for bindings (1. connect everything, 2. start sending events),
the start_X() methods on Registries needed to be added and we needed to dispatch to
different instance implementations as logic is different between artefact types (e.g. connector, pipeline, binding).

Starting work on quiescence, but this will be overhauled completely in another commit.

Signed-off-by: Matthias Wahl <mwahl@wayfair.com>
Signed-off-by: Heinz N. Gies <heinz@licenser.net>