Fix runtime bug swallowing some acks and fails
if we have an eventid tracking other streams, it used ot only return the biggest stream, not all of them.

Also adapt the cb connector to allow for testing with multiple filed forming multiple streams, so we can simulate the error condition better.

Fixed some issues with EventId, that just didn't surface yet.

Signed-off-by: Matthias Wahl <matthiaswahl@m7w3.de>